LayoutTests/fast/canvas/canvas-path-constructors-expected.txt

Test different constructors of Path.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test constructor Path().
PASS imgdata[4] is 255
PASS imgdata[5] is 255
PASS imgdata[6] is 0
PASS imgdata[7] is 255

Test constructor Path(DOMString) which takes a SVG data string.
PASS imgdata[4] is 0
PASS imgdata[5] is 0
PASS imgdata[6] is 255
PASS imgdata[7] is 255

Test constructor Path(Path) which takes another Path object.
PASS imgdata[4] is 0
PASS imgdata[5] is 128
PASS imgdata[6] is 0
PASS imgdata[7] is 255

PASS successfullyParsed is true

TEST COMPLETE
